+++ docs/examples/negotiation/validate_negotiation_settings.py
@@ -4,4 +4,5 @@ from examples.negotiation.negotiation import XmlParser, XmlRenderer
 DMR_SETTINGS = {  # noqa: WPS407
     Settings.parsers: [XmlParser()],
     Settings.renderers: [XmlRenderer()],
+    Settings.validate_negotiation: False,
 }